added linter and prettier
