+++ .eslintrc.json
@@ -1,0 +1,21 @@
+{
+	"env": {
+		"es6": true,
+		"node": true
+	},
+	"extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+	"overrides": [],
+	"parser": "@typescript-eslint/parser",
+	"parserOptions": {
+		"ecmaVersion": "latest",
+		"project": "tsconfig.json"
+	},
+	"plugins": ["@typescript-eslint"],
+	"rules": {
+		"no-var": "error",
+		"no-multi-spaces": "error",
+		"space-in-parens": "error",
+		"no-multiple-empty-lines": "error",
+		"prefer-const": "error"
+	}
+}
