+++ .prettierrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+	semi: true,
+	trailingComma: 'all',
+	singleQuote: true,
+	printWidth: 150
+};
