+++ package.json
@@ -1,29 +1,37 @@
 {
-	"name": "express-hw",
-	"version": "1.0.0",
-	"description": "Swagger basic practice",
-	"main": "src/index.js",
-	"scripts": {
-		"start": "npm run build && node build/index.js",
-		"start:dev": "npx nodemon",
-		"build": "rimraf ./build && tsc",
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"author": "Nataliia Liubynets",
-	"license": "ISC",
-	"dependencies": {
-		"express": "^4.18.2",
-		"swagger-jsdoc": "^6.2.8",
-		"swagger-ui": "^5.9.0",
-		"swagger-ui-express": "^5.0.0"
-	},
-	"devDependencies": {
-		"@types/express": "^4.17.18",
-		"@types/node": "^20.8.3",
-		"@types/swagger-ui-express": "^4.1.4",
-		"nodemon": "^3.0.1",
-		"rimraf": "^5.0.5",
-		"ts-node": "^10.9.1",
-		"typescript": "^5.2.2"
-	}
+  "name": "express-hw",
+  "version": "1.0.0",
+  "description": "Swagger basic practice",
+  "main": "src/index.js",
+  "scripts": {
+    "start": "npm run build && node build/index.js",
+    "start:dev": "npx nodemon",
+    "build": "rimraf ./build && tsc",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "eslint src/**/*.ts",
+    "format": "eslint src/**/*.ts --fix",
+    "pretty": "prettier --write \"src/**/*.ts\""
+  },
+  "author": "Nataliia Liubynets",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.18.2",
+    "prettier": "^3.0.3",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui": "^5.9.0",
+    "swagger-ui-express": "^5.0.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.18",
+    "@types/node": "^20.8.3",
+    "@types/swagger-jsdoc": "^6.0.1",
+    "@types/swagger-ui-express": "^4.1.4",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "eslint": "^8.51.0",
+    "nodemon": "^3.0.1",
+    "rimraf": "^5.0.5",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
+  }
 }
