+++ src/index.ts
@@ -1,11 +1,11 @@
-import routes from './routes/routes'; 
+import routes from './routes/routes';
 import bodyParser from 'body-parser';
 import swaggerUi from 'swagger-ui-express';
 import swaggerJsdoc from './swaggerConfig';
-import express from 'express';
+import express, { Application } from 'express';
 
-const app = express();
-const PORT = 3000;
+const app: Application = express();
+const PORT = process.env.PORT || 3000;
 
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerJsdoc));
 
@@ -13,8 +13,6 @@ app.use(bodyParser.json());
 
 app.use(routes);
 
-app.listen(PORT, () => {
-	console.log(`Server is listening on port ${PORT}`);
+app.listen(PORT, (): void => {
+  console.log(`Server is listening on port ${PORT}`);
 });
-
-
