+++ src/swaggerConfig.ts
@@ -1,0 +1,23 @@
+import swaggerJSDoc from 'swagger-jsdoc';
+
+const options = {
+  failOnErrors: true,
+  definition: {
+    openapi: '3.0.0',
+    info: {
+      title: 'SAMPLE API',
+      version: '1.0.0',
+    },
+    servers: [
+      {
+        url: 'http://localhost:3000',
+        description: 'Development server',
+      },
+    ],
+  },
+  apis: ['./src/routes/*.ts'],
+};
+
+const swaggerSpec = swaggerJSDoc(options);
+
+export default swaggerSpec;
